+++ src/views/common/Maintenance.stories.jsx
@@ -1,19 +1,30 @@
+import { configureStore } from '@reduxjs/toolkit'
 import { MaintenanceView } from '@views'
 import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
-import configureStore from 'redux-mock-store'
+
+import AppReducer from '@slices/AppSlice'
 
 export default {
   title: 'Common/MaintenanceView',
   component: MaintenanceView,
 }
 
-const mockStore = configureStore()({})
+const app = {
+  maintenance: false,
+}
+
+const store = configureStore({
+  reducer: {
+    app: AppReducer,
+  },
+  preloadedState: { app },
+})
 
 export const Default = {
   render: (props) => (
     <BrowserRouter>
-      <Provider store={mockStore}>
+      <Provider store={store}>
         <MaintenanceView {...props} />
       </Provider>
     </BrowserRouter>